testing

+++ src/main/java/com/qa/calc/Calculator.java
@@ -18,7 +18,7 @@ public class Calculator {
         return a * b;
     }
 
-    public static int divide(int a, int b) {
-        return a / b;
+    public static float divide(int a, int b) {
+        return (float) a / b;
     }
 }
